Refactor SQL Server Memory moving all SQL manipulations in external dependencies
